refactor(rmsnorm): gather the pure-torch references into rmsnorm_torch.py
rmsnorm_ref.py was introduced so ROCm could reach the RMSNorm references
without importing quack.rmsnorm, which pulls in the CuTe stack and fails
on `import cuda.bindings`. That fixed the import, but left the file as the
only *_ref.py in the tree while layernorm_ref, layernorm_rstd_ref and
layernorm_mean_ref -- equally pure torch, equally CuTe-free -- stayed
behind in quack/rmsnorm.py. "The rmsnorm references live in rmsnorm_ref.py"
was not even true of that one file.

Rename it to rmsnorm_torch.py and move the three layernorm references in,
so the module holds the whole RMSNorm/LayerNorm torch family and its name
says what it is rather than what role it plays. quack.rmsnorm re-exports
all five, so tests/test_rmsnorm.py, tests/test_layernorm.py and
benchmarks/benchmark_layernorm.py are untouched; only the shared rmsnorm
benchmark imports directly, because it is the one that must not touch CuTe.

The rsqrt comment drops its "6.7-10.9% faster for N >= 8192" claim: on an
idle gfx950 the gap measures 3-6% and shrinks with N, reaching zero by
N=32768, so the cited range does not reproduce. The justification that does
hold is structural -- rstd is what the kernels compute and return, and what
rmsnorm_bwd_ref takes as an argument -- so the comment now says that.
