+++ benchmarks/benchmark_rmsnorm.py
@@ -11,7 +11,7 @@ from triton.testing import Benchmark, do_bench, perf_report  # noqa: E402
 # quack.rmsnorm_fwd is the CuTe kernel on CUDA and the FlyDSL kernel on ROCm.
 from quack import rmsnorm_fwd  # noqa: E402
 from quack.bench.bench_utils import run_and_print  # noqa: E402
-from quack.rmsnorm_ref import rmsnorm_ref  # noqa: E402
+from quack.rmsnorm_torch import rmsnorm_ref  # noqa: E402
 
 # Inductor's donated-buffer optimization is incompatible with retain_graph=True
 # (used so we benchmark only bwd, not fwd+bwd). Disable it for the torch.compile
